chore:decomposed & query adjusted
