+++ store/store.ts
@@ -17,8 +17,6 @@ export const store = configureStore({
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({ serializableCheck: false }).concat([
       baseApi.middleware,
-      authApi.middleware,
-      inboxApi.middleware,
     ]),
 });
